fix: 🐛 change get activity details route from auth to public
